Continuing to improve testing coverage code interoperability. Added interfaces to help with both the random number generators as well as with the tick functions.

+++ src/com/company/RandomNumberGenerator.java
@@ -1,9 +1,8 @@
 package com.company;
 
 public class RandomNumberGenerator implements NumberGenerator{
-
     @Override
-    public int range(int lower, int higher) {
-        return (int) (Math.random() * (higher - lower) +lower);
+    public double range(double lower, double higher) {
+        return Math.random() * (higher - lower) +lower;
     }
 }
